+++ src/main/kotlin/de/bund/bfr/rakip/health/heatlh.kt
@@ -1,16 +1,16 @@
-package de.bund.bfr.rakip.exposuremodel
+package de.bund.bfr.rakip.health
 
 import de.bund.bfr.rakip.generic.*
 
-data class ExposureModel(
+data class ConsumptionModel(
         var generalInformation: GeneralInformation,
         var scope: Scope,
         var dataBackground: DataBackground? = null,
         var modelMath: ModelMath,
         var simulation: Simulation? = null
 )
+
 data class Scope(
-        var product: Product,
         var hazard: Hazard,
         var populationGroup: PopulationGroup,
         var generalComment: String? = null,